ARTPS/requirements.txt
# Core AI/ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Computer Vision
opencv-python>=4.5.0
pillow>=8.3.0

# Data Processing
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0

# Web Interface
streamlit>=1.28.0

# Utilities
tqdm>=4.62.0
pathlib2>=2.3.0

# Optional: For advanced features
# requests>=2.28.0  # Only if using external APIs
# beautifulsoup4>=4.11.0  # Only if web scraping
# pds4_tools>=0.7.0  # Only if using NASA PDS data 